+++ src/decompiler/decompiler.ts
@@ -2,48 +2,67 @@ import { beginCell, Cell, Slice } from "ton-core";
 import { loadOpcode } from "../codepage/loadOpcode";
 import { OpCode } from "../codepage/opcodes.gen";
 import { Maybe } from "../utils/maybe";
+import { subcell } from "../utils/subcell";
 
 export type DecompiledOpCode = OpCode | { code: 'unknown', data: Cell };
 export type DecompiledInstruction = {
     op: DecompiledOpCode,
+    hash: string,
     offset: number,
     length: number
 };
 
-export function decompile(args: { src: Cell | Slice | Buffer, srcOffset?: Maybe<number>, allowUnknown?: boolean }): DecompiledInstruction[] {
+export function decompile(args: {
+    src: Cell | Buffer,
+    offset?: Maybe<{ bits: number, refs: number }>,
+    limit?: Maybe<{ bits: number, refs: number }>,
+    allowUnknown?: boolean
+}): DecompiledInstruction[] {
 
     // Result collection
     let result: DecompiledInstruction[] = [];
 
     // Load slice
-    let sc: Slice;
+    let source: Cell;
     if (Buffer.isBuffer(args.src)) {
-        sc = Cell.fromBoc(args.src)[0].beginParse();
+        source = Cell.fromBoc(args.src)[0];
     } else if (args.src instanceof Cell) {
-        sc = args.src.beginParse();
+        source = args.src;
     } else {
-        sc = args.src;
+        throw new Error('Invalid source');
+    }
+
+    // Hash
+    let hash = source.hash().toString('hex');
+
+    // Prepare offset
+    let bitsDelta = 0;
+    let refsDelta = 0;
+    if (args.offset) {
+        bitsDelta = args.offset.bits;
+        refsDelta = args.offset.refs;
     }
 
     // Prepare offset
-    let sco = args.srcOffset || 0;
-    if (args.srcOffset && args.srcOffset > 0) {
-        sc.skip(args.srcOffset);
+    let bitsLimit = args.limit ? (args.limit.bits + bitsDelta) : source.bits.length;
+    let refsLimit = args.limit ? (args.limit.refs + refsDelta) : source.refs.length;
+    let slice = subcell({
+        cell: source,
+        bits: bitsLimit,
+        refs: refsLimit
+    }).beginParse();
+    if (args.offset) {
+        slice.skip(args.offset.bits);
+        for (let i = 0; i < args.offset.refs; i++)
+            slice.loadRef();
     }
 
-    // Prepare remaining tracker
-    let scl = sc.remainingBits;
-
-    while (sc.remainingBits > 0) {
+    while (slice.remainingBits > 0) {
 
         // Load opcode
-        const opcode = loadOpcode(sc);
-
-        // Update state
-        let currentOffset = sco; // Persisted offset before reading opcode
-        let currentLength = scl - sc.remainingBits; // Check difference in remaining bits to calculate opcode length
-        scl -= currentLength;
-        sco += currentLength;
+        const opcodeOffset = slice.offsetBits;
+        const opcode = loadOpcode(slice);
+        const opcodeLength = slice.offsetBits - opcodeOffset;
 
         // Failed case
         if (!opcode.ok) {
@@ -52,14 +71,15 @@ export function decompile(args: { src: Cell | Slice | Buffer, srcOffset?: Maybe<
                 for (let bit of Array.from(opcode.read).map(a => a == '0' ? false : true)) {
                     fullCell.storeBit(bit);
                 }
-                fullCell.storeSlice(sc);
+                fullCell.storeSlice(slice);
                 result.push({
                     op: {
                         code: 'unknown',
                         data: fullCell.endCell()
                     },
-                    offset: currentOffset,
-                    length: currentLength
+                    hash,
+                    offset: opcodeOffset,
+                    length: opcodeLength
                 });
                 break;
             } else {
@@ -70,15 +90,16 @@ export function decompile(args: { src: Cell | Slice | Buffer, srcOffset?: Maybe<
         // Push opcode to result
         result.push({
             op: opcode.read,
-            offset: currentOffset,
-            length: currentLength
+            hash,
+            offset: opcodeOffset,
+            length: opcodeLength
         });
 
         // Implicit jump
-        if (sc.remainingBits === 0 && sc.remainingRefs > 0) {
-            sc = sc.loadRef().beginParse();
-            scl = sc.remainingBits;
-            sco = 0;
+        if (slice.remainingBits === 0 && slice.remainingRefs > 0) {
+            let n = slice.loadRef();
+            hash = n.hash().toString('hex');
+            slice = n.beginParse();
         }
     }