+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Fixed
+- Fix opcode offsets and source cell hash in `PUSHCONT` opcode that could lead to broken code coverage
+
+## Changed
+- Upgrade to `ton-core@0.49.0`
+
 ## [0.0.9] - 2023-03-24
 - Fix oppcode length in shifted cells